# docs/openapi/compliance-api.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "RegEngine Compliance API",
        "description": "Multi-industry compliance checklist validation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/checklists": {
            "get": {
                "summary": "List Checklists",
                "description": "List all available compliance checklists\n\n- **industry**: Optional filter by industry (e.g., \"finance\", \"healthcare\")\n\nReturns list of checklists with metadata (ID, name, industry, jurisdiction)",
                "operationId": "list_checklists_checklists_get",
                "parameters": [
                    {
                        "name": "industry",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Industry"
                        }
                    },
                    {
                        "name": "X-RegEngine-API-Key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Regengine-Api-Key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChecklistListResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/checklists/{checklist_id}": {
            "get": {
                "summary": "Get Checklist",
                "description": "Get full details of a specific compliance checklist\n\n- **checklist_id**: Checklist identifier (e.g., \"hipaa_compliance\", \"capital_requirements\")\n\nReturns full checklist definition including all items and validation rules",
                "operationId": "get_checklist_checklists__checklist_id__get",
                "parameters": [
                    {
                        "name": "checklist_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Checklist Id"
                        }
                    },
                    {
                        "name": "X-RegEngine-API-Key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Regengine-Api-Key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/validate": {
            "post": {
                "summary": "Validate Compliance",
                "description": "Validate customer configuration against a compliance checklist\n\n**Returns yes/no pass/fail status with line-item results**\n\nRequest body:\n```json\n{\n  \"checklist_id\": \"hipaa_compliance\",\n  \"customer_config\": {\n    \"hipaa_001\": true,\n    \"hipaa_002\": true,\n    \"hipaa_003\": false\n  }\n}\n```\n\nResponse:\n```json\n{\n  \"overall_status\": \"FAIL\",\n  \"pass_rate\": 0.67,\n  \"items\": [\n    {\n      \"requirement_id\": \"hipaa_001\",\n      \"requirement\": \"Encrypt PHI at rest\",\n      \"status\": \"PASS\",\n      \"evidence\": \"Requirement met\"\n    },\n    {\n      \"requirement_id\": \"hipaa_003\",\n      \"requirement\": \"Role-Based Access Control\",\n      \"status\": \"FAIL\",\n      \"evidence\": \"Requirement not met\",\n      \"remediation\": \"Implement RBAC system...\"\n    }\n  ],\n  \"next_steps\": [\n    \"Address failed requirements before launching\",\n    \"\u2192 Implement RBAC for PHI access\"\n  ]\n}\n```",
                "operationId": "validate_compliance_validate_post",
                "parameters": [
                    {
                        "name": "X-RegEngine-API-Key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Regengine-Api-Key"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ValidationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/industries": {
            "get": {
                "summary": "List Industries",
                "description": "List all supported industries\n\nReturns: List of industry names with checklist counts",
                "operationId": "list_industries_industries_get",
                "parameters": [
                    {
                        "name": "X-RegEngine-API-Key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Regengine-Api-Key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/fsma-204/assess": {
            "post": {
                "summary": "Assess Fsma Readiness",
                "description": "Run a FSMA 204 readiness assessment",
                "operationId": "assess_fsma_readiness_fsma_204_assess_post",
                "parameters": [
                    {
                        "name": "X-RegEngine-API-Key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Regengine-Api-Key"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FSMAAssessmentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FSMAAssessmentResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/fsma/audit/spreadsheet": {
            "get": {
                "summary": "Get Fsma Audit Spreadsheet",
                "description": "Generate FDA-compliant CSV spreadsheet for FSMA 204 audit.\n\nArgs:\n    tlc: Traceability Lot Code to trace\n    start_date: Start date filter (YYYY-MM-DD)\n    end_date: End date filter (YYYY-MM-DD)\n    \nReturns:\n    Downloadable CSV file",
                "operationId": "get_fsma_audit_spreadsheet_fsma_audit_spreadsheet_get",
                "parameters": [
                    {
                        "name": "tlc",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Tlc"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Start Date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "End Date"
                        }
                    },
                    {
                        "name": "X-RegEngine-API-Key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Regengine-Api-Key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/fsma/validate-tlc": {
            "post": {
                "summary": "Validate Tlc",
                "description": "Validate a Traceability Lot Code format.\n\nArgs:\n    tlc: Traceability Lot Code to validate\n    \nReturns:\n    Validation result with error message if invalid",
                "operationId": "validate_tlc_fsma_validate_tlc_post",
                "parameters": [
                    {
                        "name": "tlc",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Tlc"
                        }
                    },
                    {
                        "name": "X-RegEngine-API-Key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Regengine-Api-Key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ChecklistListResponse": {
                "properties": {
                    "checklists": {
                        "items": {
                            "type": "object"
                        },
                        "type": "array",
                        "title": "Checklists"
                    },
                    "total": {
                        "type": "integer",
                        "title": "Total"
                    }
                },
                "type": "object",
                "required": [
                    "checklists",
                    "total"
                ],
                "title": "ChecklistListResponse",
                "description": "List of available compliance checklists"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
            "ValidationRequest": {
                "properties": {
                    "checklist_id": {
                        "type": "string",
                        "title": "Checklist Id",
                        "description": "ID of the compliance checklist"
                    },
                    "customer_config": {
                        "type": "object",
                        "title": "Customer Config",
                        "description": "Customer configuration/answers keyed by requirement ID",
                        "example": {
                            "hipaa_001": true,
                            "hipaa_002": true,
                            "hipaa_003": false
                        }
                    }
                },
                "type": "object",
                "required": [
                    "checklist_id",
                    "customer_config"
                ],
                "title": "ValidationRequest",
                "description": "Request to validate compliance against a checklist"
            },
            "ValidationResponse": {
                "properties": {
                    "checklist_id": {
                        "type": "string",
                        "title": "Checklist Id"
                    },
                    "checklist_name": {
                        "type": "string",
                        "title": "Checklist Name"
                    },
                    "industry": {
                        "type": "string",
                        "title": "Industry"
                    },
                    "jurisdiction": {
                        "type": "string",
                        "title": "Jurisdiction"
                    },
                    "overall_status": {
                        "type": "string",
                        "title": "Overall Status"
                    },
                    "pass_rate": {
                        "type": "number",
                        "title": "Pass Rate",
                        "description": "Pass rate as decimal (0.0-1.0)"
                    },
                    "items": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationItemResponse"
                        },
                        "type": "array",
                        "title": "Items"
                    },
                    "next_steps": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Next Steps"
                    }
                },
                "type": "object",
                "required": [
                    "checklist_id",
                    "checklist_name",
                    "industry",
                    "jurisdiction",
                    "overall_status",
                    "pass_rate",
                    "items",
                    "next_steps"
                ],
                "title": "ValidationResponse",
                "description": "Complete validation result"
            }
        }
    }
}